+++ src/roon-integration.ts
@@ -19,11 +19,11 @@ import os from "os";
 
 export default class RoonDriver {
   private driver: uc.IntegrationAPI;
-  private roon: RoonApi;
-  private roonApiStatus: RoonApiStatus;
+  private readonly roon: RoonApi;
+  private readonly roonApiStatus: RoonApiStatus;
   private roonCore: Core | null = null;
   private roonImage: RoonApiImage | null = null;
-  private roonTransport: RoonApiTransport;
+  private roonTransport: RoonApiTransport | null = null;
   private roonPaired = false;
   private config: Config;
 
@@ -140,7 +140,7 @@ export default class RoonDriver {
         // update entity with current Zone information
         const zone = this.roonTransport?.zone_by_zone_id(entityId);
         if (zone) {
-          log.info(`Zone data: ${JSON.stringify(zone)}`);
+          log.debug(`Zone data: ${JSON.stringify(zone)}`);
           const attr = mediaPlayerAttributesFromZone(zone);
           this.driver.getConfiguredEntities().updateEntityAttributes(entity.id, attr);
         } else {
@@ -153,19 +153,21 @@ export default class RoonDriver {
 
   private async handleDisconnect() {
     this.roonApiStatus?.set_status("Disconnected", false);
-    // TODO unsubscribe from Roon?
+    // TODO unsubscribe from Roon? how?
     await this.driver.setDeviceState(uc.DeviceStates.Disconnected);
   }
 
   private async handleEnterStandby() {
     log.debug("enter standby");
     this.roonApiStatus?.set_status("Standby", false);
-    // TODO #56 force Roon disconnect
+    // TODO #56 force Roon disconnect, but how?
   }
 
   private async handleExitStandby() {
     log.debug("exit standby");
     // TODO #56 force Roon reconnect
+    // No disconnect method found in API! Let's try sending a request, maybe this triggers a reconnect
+    this.roonTransport?.get_zones();
     this.roonApiStatus?.set_status("Connected", false);
   }
 
@@ -182,9 +184,24 @@ export default class RoonDriver {
 
     // add event listeners to roon
     this.roonTransport.subscribe_zones(async (cmd, msg) => {
+      // log.debug("subscribe_zones callback: %s - %s", cmd, JSON.stringify(msg))
       switch (cmd) {
         case "Changed": {
           const data = msg as SubscribeZoneChanged;
+          if (data.zones_added) {
+            data.zones_added.forEach((zone: Zone) => {
+              log.info(`Zone added: ${zone.display_name}`);
+              this.updateMediaPlayerFromZone(zone);
+            });
+          }
+
+          if (data.zones_removed) {
+            data.zones_removed.forEach((zone_id) => {
+              log.info(`Zone removed: ${zone_id}`);
+              this.setEntityState(zone_id, uc.MediaPlayerStates.Unavailable);
+            });
+          }
+
           if (data.zones_changed) {
             data.zones_changed.forEach((zone: Zone) => {
               log.debug(`Zone changed: ${zone.display_name}`);
@@ -192,6 +209,10 @@ export default class RoonDriver {
             });
           } else if (data.zones_seek_changed) {
             data.zones_seek_changed.forEach((zone) => {
+              // ignore zone if it's not in our zone configuration
+              if (!this.config.hasZone(zone.zone_id)) {
+                return;
+              }
               if (!this.driver.getConfiguredEntities().contains(zone.zone_id)) {
                 log.debug(`Configured entity not found, not updating seek:(${zone.zone_id})`);
                 return;
@@ -215,6 +236,10 @@ export default class RoonDriver {
           }
           break;
         }
+        case "Unsubscribed": {
+          log.debug("Zone unsubscribe %s", JSON.stringify(msg));
+          break;
+        }
       }
     });
   }
@@ -228,6 +253,11 @@ export default class RoonDriver {
     if (!zone) {
       return false;
     }
+    // ignore zone if it's not in our zone configuration
+    if (!this.config.hasZone(zone.zone_id)) {
+      return false;
+    }
+
     if (!this.driver.getConfiguredEntities().contains(zone.zone_id)) {
       log.info(`Configured entity not found, not updating: ${zone.display_name} (${zone.zone_id})`);
       return false;
@@ -246,7 +276,7 @@ export default class RoonDriver {
             height: 480,
             format: "image/jpeg"
           },
-          (error, contentType, image) => {
+          (error, _contentType, image) => {
             if (error) {
               log.warn(`Failed to get image: ${error}`);
             } else if (image) {
@@ -277,6 +307,7 @@ export default class RoonDriver {
     this.roonPaired = false;
     this.roonCore = null;
     this.roonImage = null;
+    this.roonTransport = null;
     // #56 set all entities to unavailable if we are no longer paired with the Roon core
     // TODO enhance integration-library with helper functions to retrieve all entity IDs or update all entity states
     this.driver
@@ -292,17 +323,17 @@ export default class RoonDriver {
 
   private async getRoonZones(): Promise<void> {
     if (this.roonCore == null) {
-      log.warn("Cannot get Roon zones. RoonCore is null.");
+      log.warn("Cannot get Roon zones. Roon Core not available.");
       return;
     }
 
     if (this.roonTransport == null) {
-      log.warn("Cannot get Roon zones. RoonTransport is null.");
+      log.warn("Cannot get Roon zones. Roon Transport not available.");
       return;
     }
 
     return new Promise((resolve, reject) => {
-      this.roonTransport.get_zones(async (error, data) => {
+      this.roonTransport?.get_zones(async (error, data) => {
         if (error) {
           log.warn("Failed to get Roon Zones");
           reject(error);
@@ -348,9 +379,10 @@ export default class RoonDriver {
         case uc.MediaPlayerCommands.PlayPause: {
           const roonCmd =
             entity?.attributes?.[uc.MediaPlayerAttributes.State] === uc.MediaPlayerStates.Playing ? "pause" : "play";
-          this.roonTransport.control(entity.id, roonCmd, async (error) => {
+          this.roonTransport?.control(entity.id, roonCmd, async (error) => {
             if (error) {
               log.error(`Error on ${roonCmd} media player: ${error}`);
+              // TODO parse `error` and handle certain error conditions like `ZoneNotFound`
               resolve(uc.StatusCodes.ServerError);
             } else {
               resolve(uc.StatusCodes.Ok);
@@ -359,7 +391,7 @@ export default class RoonDriver {
           break;
         }
         case uc.MediaPlayerCommands.Next:
-          this.roonTransport.control(entity.id, "next", async (error) => {
+          this.roonTransport?.control(entity.id, "next", async (error) => {
             if (error) {
               log.error(`Error next media player: ${error}`);
               resolve(uc.StatusCodes.ServerError);
@@ -369,7 +401,7 @@ export default class RoonDriver {
           });
           break;
         case uc.MediaPlayerCommands.Previous:
-          this.roonTransport.control(entity.id, "previous", async (error) => {
+          this.roonTransport?.control(entity.id, "previous", async (error) => {
             if (error) {
               log.error(`Error previous media player: ${error}`);
               resolve(uc.StatusCodes.ServerError);
@@ -381,7 +413,7 @@ export default class RoonDriver {
         case uc.MediaPlayerCommands.Volume: {
           const output = this.getDefaultZoneOutput(entity.id);
           if (output) {
-            this.roonTransport.change_volume(output.output_id, "absolute", Number(params?.volume), async (error) => {
+            this.roonTransport?.change_volume(output.output_id, "absolute", Number(params?.volume), async (error) => {
               if (error) {
                 log.error(`Error changing volume media player: ${error}`);
                 resolve(uc.StatusCodes.ServerError);
@@ -395,7 +427,7 @@ export default class RoonDriver {
         case uc.MediaPlayerCommands.VolumeUp: {
           const output = this.getDefaultZoneOutput(entity.id);
           if (output) {
-            this.roonTransport.change_volume(output.output_id, "relative_step", 1, async (error) => {
+            this.roonTransport?.change_volume(output.output_id, "relative_step", 1, async (error) => {
               if (error) {
                 log.error(`Error changing volume media player: ${error}`);
                 resolve(uc.StatusCodes.ServerError);
@@ -412,7 +444,7 @@ export default class RoonDriver {
         case uc.MediaPlayerCommands.VolumeDown: {
           const output = this.getDefaultZoneOutput(entity.id);
           if (output) {
-            this.roonTransport.change_volume(output.output_id, "relative_step", -1, async (error) => {
+            this.roonTransport?.change_volume(output.output_id, "relative_step", -1, async (error) => {
               if (error) {
                 log.error(`Error changing volume media player: ${error}`);
                 resolve(uc.StatusCodes.ServerError);
@@ -427,7 +459,7 @@ export default class RoonDriver {
           const output = this.getDefaultZoneOutput(entity.id);
           if (output) {
             const roonCmd = entity.attributes?.[uc.MediaPlayerAttributes.Muted] ? "unmute" : "mute";
-            this.roonTransport.mute(output.output_id, roonCmd, async (error) => {
+            this.roonTransport?.mute(output.output_id, roonCmd, async (error) => {
               if (error) {
                 log.error(`Error on ${roonCmd} media player: ${error}`);
                 resolve(uc.StatusCodes.ServerError);
@@ -439,7 +471,7 @@ export default class RoonDriver {
           break;
         }
         case uc.MediaPlayerCommands.Seek:
-          this.roonTransport.seek(entity.id, "absolute", Number(params?.media_position), async (error) => {
+          this.roonTransport?.seek(entity.id, "absolute", Number(params?.media_position), async (error) => {
             if (error) {
               log.error(`Error seeking media player: ${error}`);
               resolve(uc.StatusCodes.ServerError);
@@ -460,6 +492,10 @@ export default class RoonDriver {
   }
 
   private setEntityState(entityId: string, state: uc.MediaPlayerStates) {
+    // ignore zone if it's not in our zone configuration
+    if (!this.config.hasZone(entityId)) {
+      return;
+    }
     this.driver.getConfiguredEntities().updateEntityAttributes(entityId, {
       [uc.MediaPlayerAttributes.State]: state
     });